split managers and resources
