+++ src/managers/bms.rs
@@ -1,11 +1,12 @@
 // All tasks operating on the BMS (Battery Management System) live in this file
 use crate::managers::Manager;
 use crate::managers::ManagerChannelData;
-use crate::managers::Resource;
 use crate::managers::ResourceManager;
 use crate::managers::TASK_SUCCESS;
 use crate::managers::macros::parse_channel_data;
 use crate::request::TaskData::BmsData;
+use crate::resources::Resource;
+use crate::resources::bms::Bms;
 use crate::sgcp;
 use crate::sgcp::bms::*;
 use crate::todo;
@@ -13,21 +14,6 @@ use anyhow::Error;
 use anyhow::Result;
 use anyhow::anyhow;
 use log::*;
-
-/// Represents a BMS resource
-pub struct Bms {
-    // TODO: @krarpit Implement BMS interface
-}
-
-impl Resource for Bms {
-    fn init() -> Self {
-        Bms {} // stub
-    }
-
-    fn name() -> String {
-        sgcp::Resource::Bms.as_str_name().to_string()
-    }
-}
 
 impl ResourceManager for Manager<Bms> {
     type ResourceType = Bms;
@@ -44,3 +30,4 @@ impl ResourceManager for Manager<Bms> {
             .map_err(|e| anyhow!("Send Failed: {e}"))?)
     }
 }
+
