+++ src/managers/emg/adc.rs
@@ -1,8 +1,8 @@
-use anyhow::{Result, Error, Context};
+use anyhow::{Context, Error, Result};
 use log::*;
 
-use rppal::gpio::{OutputPin};
-use rppal::spi::{Spi};
+use rppal::gpio::OutputPin;
+use rppal::spi::Spi;
 use std::{io, thread, time::Duration};
 
 pub fn process_data(values: Vec<u16>, inner_threshold: u16, outer_threshold: u16) -> Result<i32> {
@@ -10,12 +10,11 @@ pub fn process_data(values: Vec<u16>, inner_threshold: u16, outer_threshold: u16
         return Err(Error::msg("Expected 2 EMG values"));
     }
 
-    if values[0] >= inner_threshold && values[1] <= outer_threshold{
+    if values[0] >= inner_threshold && values[1] <= outer_threshold {
         Ok(1) // Open
     } else if values[0] <= inner_threshold && values[1] >= outer_threshold {
         Ok(0) // Close
-    }
-    else {
+    } else {
         Ok(-1) // No action
     }
 }
@@ -28,8 +27,13 @@ fn average(list: Vec<u16>) -> Result<u16> {
         Ok((sum / list.len() as u32) as u16)
     }
 }
-    
-pub fn calibrate_emg(buffer_size : usize, spi: &mut Spi, cs: &mut OutputPin, pause_duration : u64) -> Result<[u16; 2]> {
+
+pub fn calibrate_emg(
+    buffer_size: usize,
+    spi: &mut Spi,
+    cs: &mut OutputPin,
+    pause_duration: u64,
+) -> Result<[u16; 2]> {
     let inner_buffer = read_samples(0, cs, buffer_size, spi, "inner", pause_duration);
 
     info!("\nFinished inner sampling. Press ENTER when you're ready to start outer sampling...");
@@ -51,7 +55,14 @@ pub fn calibrate_emg(buffer_size : usize, spi: &mut Spi, cs: &mut OutputPin, pau
     Ok([avg_inner, avg_outer])
 }
 
-pub fn read_samples(channel: u8, cs: &mut OutputPin, sample_count: usize, spi: &mut Spi, label: &str, pause_duration : u64) -> Vec<u16> {
+pub fn read_samples(
+    channel: u8,
+    cs: &mut OutputPin,
+    sample_count: usize,
+    spi: &mut Spi,
+    label: &str,
+    pause_duration: u64,
+) -> Vec<u16> {
     let mut buffer = Vec::with_capacity(sample_count);
     info!("Flex {label}");
 
@@ -65,7 +76,6 @@ pub fn read_samples(channel: u8, cs: &mut OutputPin, sample_count: usize, spi: &
 
     buffer
 }
-
 
 pub fn read_adc_channels(channels: &[u8], cs: &mut OutputPin, spi: &mut Spi) -> Result<Vec<u16>> {
     channels
@@ -82,7 +92,10 @@ pub fn read_adc_channels(channels: &[u8], cs: &mut OutputPin, spi: &mut Spi) -> 
 /// doc link: https://www.mathworks.com/help/matlab/supportpkg/analog-input-using-spi.html
 pub fn read_adc_channel(channel: u8, cs: &mut OutputPin, spi: &mut Spi) -> Result<u16> {
     if channel > 7 {
-        return Err(Error::msg(format!("Invalid ADC channel: {}. Must be between 0 and 7.", channel)));
+        return Err(Error::msg(format!(
+            "Invalid ADC channel: {}. Must be between 0 and 7.",
+            channel
+        )));
     }
 
     // first byte: start sequence, sends 1 as start bit
@@ -93,14 +106,14 @@ pub fn read_adc_channel(channel: u8, cs: &mut OutputPin, spi: &mut Spi) -> Resul
     // then combines with bitwise OR operator
     // remaining bits are ignored
     let config_bits = 0b10000000 | (channel << 4);
-    
+
     // third byte: dummy to clock out rest of ADC result
     let tx = [start_bit, config_bits, 0x00];
 
     // response buffer
     let mut rx = [0u8; 3];
 
-    //activate chip select 
+    //activate chip select
     cs.set_low();
 
     // full-duplex SPI transfer: sends tx[], fills rx[]
@@ -110,10 +123,10 @@ pub fn read_adc_channel(channel: u8, cs: &mut OutputPin, spi: &mut Spi) -> Resul
     // deactivate chip select
     cs.set_high();
 
-
     // adc sends back 3 byte response
     // actual response is 10 bits and spread across rx[1](bits 9-8) and rx[2](bits 7-0)
     // isolate rx[1] result bits, then shift them to 9-8 in a 16 bit number, then add remaining bits by combining them with bitwise OR
     let result = ((rx[1] & 0b00000011) as u16) << 8 | (rx[2] as u16);
     Ok(result)
 }
+
