--- src/managers/resources/mod.rs
@@ -1,3 +1,0 @@
-pub mod bms;
-pub mod maestro;
-pub mod emg;
