+++ src/resources/mod.rs
@@ -1,0 +1,8 @@
+pub mod bms;
+pub mod emg;
+pub mod maestro;
+
+pub trait Resource {
+    fn init() -> Self;
+    fn name() -> String;
+}
